tests/smoke/examples/synthetic/flask_azure_blob/requirements.txt
flask==3.0.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
python-dotenv==1.0.0
